Fix test and add a new one

+++ src/main/kotlin/ru/hse/smartUniversity/UniversityMap.kt
@@ -3,11 +3,11 @@ package ru.hse.smartUniversity
 class UniversityMap {
     private data class Room(val name: String) {
         val neighbors = mutableSetOf<Room>()
-        val distances = mutableMapOf<Room, Int>()
+        val distances = mutableMapOf<Room, Long>()
     }
 
     private val rooms = mutableMapOf<String, Room>()
-    private fun connect(first: Room?, second: Room?, dist: Int) {
+    private fun connect(first: Room?, second: Room?, dist: Long) {
         if (second != null) first?.neighbors?.add(second)
         if (first != null) second?.neighbors?.add(first)
         if (second != null) first?.distances?.put(second, dist)
@@ -18,13 +18,13 @@ class UniversityMap {
         rooms[name] = Room(name)
     }
 
-    fun connect(first: String, second: String, dist: Int = 1) =
+    fun connect(first: String, second: String, dist: Long = 1) =
         connect(rooms[first], rooms[second], dist)
 
     fun neighbors(name: String): List<String> =
         rooms[name]?.neighbors?.map { it.name } ?: listOf()
 
-    fun getDist(from: String, to: String): Int =
+    fun getDist(from: String, to: String): Long =
         rooms[from]?.distances?.get(rooms[to]) ?: -1
 
     fun getAllRooms(): List<String> =
